+++ packages/composites/outlier-detection/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/outlier-detection-adapter
 
+## 1.0.51
+
+### Patch Changes
+
+- Updated dependencies [d0b872f6c]
+  - @chainlink/ea-bootstrap@1.12.0
+  - @chainlink/ea-reference-data-reader@1.1.6
+  - @chainlink/ea-test-helpers@1.1.0
+  - @chainlink/ea@1.0.51
+
 ## 1.0.50
 
 ### Patch Changes