+++ packages/composites/the-graph/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/the-graph-adapter
 
+## 1.0.31
+
+### Patch Changes
+
+- Updated dependencies [d0b872f6c]
+  - @chainlink/ea-bootstrap@1.12.0
+  - @chainlink/ea-reference-data-reader@1.1.6
+  - @chainlink/ea-test-helpers@1.1.0
+  - @chainlink/graphql-adapter@1.1.16
+
 ## 1.0.30
 
 ### Patch Changes