+++ packages/core/reference-data-reader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/ea-reference-data-reader
 
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [d0b872f6c]
+  - @chainlink/ea-bootstrap@1.12.0
+
 ## 1.1.5
 
 ### Patch Changes