+++ packages/sources/oilpriceapi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/oilpriceapi-adapter
 
+## 2.0.16
+
+### Patch Changes
+
+- Updated dependencies [d0b872f6c]
+  - @chainlink/ea-bootstrap@1.12.0
+  - @chainlink/ea-test-helpers@1.1.0
+
 ## 2.0.15
 
 ### Patch Changes